Further folder structure reorg/consolidation
