--- apps/react/demo/src/ui/Example/Provider/index.ts
@@ -1,2 +1,0 @@
-export { default as Provider } from "./Provider.js";
-export * from "./types.js";
