+++ apps/react/demo/src/ui/Showcase/common/Example/common/examples/TypographicSpecimen/index.ts
@@ -1,2 +1,1 @@
 export { default as TypographicSpecimen } from "./TypographicSpecimen.js";
-export * from "./types.js";
