+++ apps/react/demo/src/ui/Showcase/common/Example/hooks/types.ts
@@ -1,0 +1,5 @@
+import type { ContextOptions, ProviderProps } from "../types.js";
+
+export type UseProviderStateProps = Omit<ProviderProps, "children">;
+
+export type UseProviderStateResult = ContextOptions;
