+++ apps/react/demo/src/ui/Showcase/common/Example/hooks/useConfig.ts
