+++ apps/react/demo/src/ui/Showcase/common/Example/index.ts
@@ -2,7 +2,7 @@ import type { ExampleComponent } from "./types.js";
 
 export * from "./types.js";
 
-import { Provider } from "./Provider/index.js";
+import Provider from "./Provider.js";
 import { Controls } from "./common/Controls/index.js";
 import { Renderer } from "./common/Renderer/index.js";
 export const Example = Provider as ExampleComponent;
