+++ apps/react/demo/src/ui/Showcase/common/index.ts
@@ -1,0 +1,1 @@
+export * from "./Example/index.js";
