+++ apps/react/demo/src/ui/Showcase/index.ts
@@ -1,1 +1,2 @@
 export { default as Showcase } from "./Showcase.js";
+export * from "./common/index.js";
